Refactoring: moving DomainConfigurationAware trait to controllers package
